:recycle: Refactoring categoryRepository without middleware
